backend/requirements.txt
django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
mssql-django==1.5
pyodbc==5.1.0
python-dotenv==1.0.1
gunicorn==22.0.0